feat: Stage 2 - Extract Pydantic models to separate module
- Created models/schemas.py with all Pydantic models
- Moved 5 models: SlackTokenValidation, ChannelInfo, MessageInfo, UserInfo, FileUploadInfo
- Updated models/__init__.py to export all models
- Updated slack_mcp_server.py to import models from new location
- Removed original model definitions from main server file
- Maintained exact field definitions and descriptions for MCP compatibility
- Updated README.md to document Stage 2 completion

All functionality preserved with improved code organization.
